Move extensions logic to Extensions section component
