+++ assets/home/main.js
@@ -1,11 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { Notice, Spinner } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
-import { useSelect } from '@wordpress/data';
 import { EditorNotices } from '@wordpress/editor';
-import { RawHTML } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -13,7 +9,6 @@ import { RawHTML } from '@wordpress/element';
 import { useSenseiColorTheme } from '../react-hooks/use-sensei-color-theme';
 import FeaturedProductSenseiPro from './featured-product-sensei-pro';
 import Header from './header';
-import { EXTENSIONS_STORE } from '../extensions/store';
 import { Col, Grid } from './grid';
 import QuickLinks from './sections/quick-links';
 import TaskList from './sections/task-list';
@@ -25,45 +20,11 @@ import Extensions from './sections/extensions';
 const Main = () => {
 	useSenseiColorTheme();
 
-	const { extensions, layout, isExtensionsLoading, error } = useSelect(
-		( select ) => {
-			const store = select( EXTENSIONS_STORE );
-
-			return {
-				isExtensionsLoading: ! store.hasFinishedResolution(
-					'getExtensions'
-				),
-				extensions: store.getExtensions(),
-				layout: store.getLayout(),
-				error: store.getError(),
-			};
-		},
-		[]
-	);
-
-	if ( isExtensionsLoading ) {
-		return (
-			<div className="sensei-home__loader">
-				<Spinner />
-			</div>
-		);
-	}
-
-	if ( 0 === extensions.length || 0 === layout.length ) {
-		return <div>{ __( 'No extensions found.', 'sensei-lms' ) }</div>;
-	}
-
 	return (
 		<>
 			<Grid as="main" className="sensei-home">
 				<Col className="sensei-home__section" cols={ 12 }>
 					<Header />
-
-					{ error !== null && (
-						<Notice status="error" isDismissible={ false }>
-							<RawHTML>{ error }</RawHTML>
-						</Notice>
-					) }
 				</Col>
 
 				<Col cols={ 12 }>
@@ -91,7 +52,7 @@ const Main = () => {
 				</Col>
 
 				<Col cols={ 12 }>
-					<Extensions extensions={ extensions } />
+					<Extensions />
 				</Col>
 			</Grid>
 			<EditorNotices />
